feat: crate refactor
