+++ crates/derive/src/traits/data_sources.rs
@@ -1,55 +1,15 @@
 //! Contains traits that describe the functionality of various data sources used in the derivation
 //! pipeline's stages.
 
-use crate::types::{
-    Blob, BlockInfo, IndexedBlobHash, L2BlockInfo, L2ExecutionPayloadEnvelope, RollupConfig,
-    StageResult, SystemConfig,
-};
+use crate::errors::StageResult;
 use alloc::{boxed::Box, fmt::Debug, sync::Arc, vec::Vec};
 use alloy_consensus::{Header, Receipt, TxEnvelope};
 use alloy_primitives::{Address, Bytes, B256};
 use anyhow::Result;
 use async_trait::async_trait;
-
-/// Describes the functionality of a data source that can provide information from the blockchain.
-#[async_trait]
-pub trait ChainProvider {
-    /// Fetch the L1 [Header] for the given [B256] hash.
-    async fn header_by_hash(&mut self, hash: B256) -> Result<Header>;
-
-    /// Returns the block at the given number, or an error if the block does not exist in the data
-    /// source.
-    async fn block_info_by_number(&mut self, number: u64) -> Result<BlockInfo>;
-
-    /// Returns all receipts in the block with the given hash, or an error if the block does not
-    /// exist in the data source.
-    async fn receipts_by_hash(&mut self, hash: B256) -> Result<Vec<Receipt>>;
-
-    /// Returns the [BlockInfo] and list of [TxEnvelope]s from the given block hash.
-    async fn block_info_and_transactions_by_hash(
-        &mut self,
-        hash: B256,
-    ) -> Result<(BlockInfo, Vec<TxEnvelope>)>;
-}
-
-/// Describes the functionality of a data source that fetches safe blocks.
-#[async_trait]
-pub trait L2ChainProvider {
-    /// Returns the L2 block info given a block number.
-    /// Errors if the block does not exist.
-    async fn l2_block_info_by_number(&mut self, number: u64) -> Result<L2BlockInfo>;
-
-    /// Returns an execution payload for a given number.
-    /// Errors if the execution payload does not exist.
-    async fn payload_by_number(&mut self, number: u64) -> Result<L2ExecutionPayloadEnvelope>;
-
-    /// Returns the [SystemConfig] by L2 number.
-    async fn system_config_by_number(
-        &mut self,
-        number: u64,
-        rollup_config: Arc<RollupConfig>,
-    ) -> Result<SystemConfig>;
-}
+use kona_primitives::block::BlockInfo;
+use kona_primitives::blob::{Blob, IndexedBlobHash};
+pub use kona_providers::prelude::{ChainProvider, L2ChainProvider};
 
 /// The BlobProvider trait specifies the functionality of a data source that can provide blobs.
 #[async_trait]
