+++ crates/derive/src/traits/stages.rs
@@ -1,8 +1,11 @@
 //! This module contains common traits for stages within the derivation pipeline.
 
-use crate::types::{BlockInfo, StageResult, SystemConfig};
+use kona_primitives::block::BlockInfo;
+use kona_primitives::system_config::SystemConfig;
 use alloc::boxed::Box;
 use async_trait::async_trait;
+
+use crate::errors::StageResult;
 
 /// Describes the functionality fo a resettable stage within the derivation pipeline.
 #[async_trait]
