+++ crates/primitives/src/batch/span_batch/errors.rs
