+++ crates/primitives/src/batch/span_batch/raw.rs
