+++ crates/primitives/src/batch/span_batch/signature.rs
