+++ crates/primitives/src/batch/span_batch/utils.rs
