+++ crates/primitives/src/batch/validity.rs
