+++ crates/primitives/src/deposits.rs
@@ -1,14 +1,162 @@
 //! Contains deposit transaction types and helper methods.
 
-use alloc::{string::String, vec::Vec};
-use alloy_primitives::{keccak256, Address, Bytes, Log, TxKind, B256, U256, U64};
+use crate::raw_tx::RawTransaction;
+use alloc::{fmt::Display, string::String, vec::Vec};
+use alloy_primitives::{b256, keccak256, Address, Bytes, Log, TxKind, B256, U256, U64};
 use alloy_rlp::Encodable;
 use op_alloy_consensus::TxDeposit;
 
-use crate::{
-    params::DEPOSIT_EVENT_ABI_HASH,
-    types::{DepositError, RawTransaction},
-};
+/// Deposit log event abi signature.
+pub const DEPOSIT_EVENT_ABI: &str = "TransactionDeposited(address,address,uint256,bytes)";
+
+/// Deposit event abi hash.
+///
+/// This is the keccak256 hash of the deposit event ABI signature.
+/// `keccak256("TransactionDeposited(address,address,uint256,bytes)")`
+pub const DEPOSIT_EVENT_ABI_HASH: B256 =
+    b256!("b3813568d9991fc951961fcb4c784893574240a28925604d09fc577c55bb7c32");
+
+/// The initial version of the deposit event log.
+pub const DEPOSIT_EVENT_VERSION_0: B256 = B256::ZERO;
+
+/// An [op_alloy_consensus::TxDeposit] validation error.
+#[derive(Debug)]
+pub enum DepositError {
+    /// Unexpected number of deposit event log topics.
+    UnexpectedTopicsLen(usize),
+    /// Invalid deposit event selector.
+    /// Expected: [B256] (deposit event selector), Actual: [B256] (event log topic).
+    InvalidSelector(B256, B256),
+    /// Incomplete opaqueData slice header (incomplete length).
+    IncompleteOpaqueData(usize),
+    /// The log data is not aligned to 32 bytes.
+    UnalignedData(usize),
+    /// Failed to decode the `from` field of the deposit event (the second topic).
+    FromDecode(B256),
+    /// Failed to decode the `to` field of the deposit event (the third topic).
+    ToDecode(B256),
+    /// Invalid opaque data content offset.
+    InvalidOpaqueDataOffset(Bytes),
+    /// Invalid opaque data content length.
+    InvalidOpaqueDataLength(Bytes),
+    /// Opaque data length exceeds the deposit log event data length.
+    /// Specified: [usize] (data length), Actual: [usize] (opaque data length).
+    OpaqueDataOverflow(usize, usize),
+    /// Opaque data with padding exceeds the specified data length.
+    PaddedOpaqueDataOverflow(usize, usize),
+    /// An invalid deposit version.
+    InvalidVersion(B256),
+    /// Unexpected opaque data length
+    UnexpectedOpaqueDataLen(usize),
+    /// Failed to decode the deposit mint value.
+    MintDecode(Bytes),
+    /// Failed to decode the deposit gas value.
+    GasDecode(Bytes),
+    /// A custom error wrapping [anyhow::Error].
+    Custom(anyhow::Error),
+}
+
+impl PartialEq<DepositError> for DepositError {
+    fn eq(&self, other: &DepositError) -> bool {
+        match (self, other) {
+            (DepositError::UnexpectedTopicsLen(l1), DepositError::UnexpectedTopicsLen(l2)) => {
+                l1 == l2
+            }
+            (DepositError::InvalidSelector(e1, t1), DepositError::InvalidSelector(e2, t2)) => {
+                e1 == e2 && t1 == t2
+            }
+            (DepositError::IncompleteOpaqueData(l1), DepositError::IncompleteOpaqueData(l2)) => {
+                l1 == l2
+            }
+            (DepositError::UnalignedData(d1), DepositError::UnalignedData(d2)) => d1 == d2,
+            (DepositError::FromDecode(e1), DepositError::FromDecode(e2)) => e1 == e2,
+            (DepositError::ToDecode(e1), DepositError::ToDecode(e2)) => e1 == e2,
+            (
+                DepositError::InvalidOpaqueDataOffset(o1),
+                DepositError::InvalidOpaqueDataOffset(o2),
+            ) => o1 == o2,
+            (
+                DepositError::InvalidOpaqueDataLength(o1),
+                DepositError::InvalidOpaqueDataLength(o2),
+            ) => o1 == o2,
+            (
+                DepositError::OpaqueDataOverflow(l1, l2),
+                DepositError::OpaqueDataOverflow(l3, l4),
+            ) => l1 == l3 && l2 == l4,
+            (
+                DepositError::PaddedOpaqueDataOverflow(l1, l2),
+                DepositError::PaddedOpaqueDataOverflow(l3, l4),
+            ) => l1 == l3 && l2 == l4,
+            (DepositError::InvalidVersion(v1), DepositError::InvalidVersion(v2)) => v1 == v2,
+            (
+                DepositError::UnexpectedOpaqueDataLen(a),
+                DepositError::UnexpectedOpaqueDataLen(b),
+            ) => a == b,
+            (DepositError::MintDecode(a), DepositError::MintDecode(b)) => a == b,
+            (DepositError::GasDecode(a), DepositError::GasDecode(b)) => a == b,
+            (DepositError::Custom(_), DepositError::Custom(_)) => true,
+            _ => false,
+        }
+    }
+}
+
+impl Display for DepositError {
+    fn fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
+        match self {
+            DepositError::UnexpectedTopicsLen(len) => {
+                write!(f, "Unexpected number of deposit event log topics: {}", len)
+            }
+            DepositError::InvalidSelector(expected, actual) => {
+                write!(f, "Invalid deposit event selector: {}, expected {}", actual, expected)
+            }
+            DepositError::IncompleteOpaqueData(len) => {
+                write!(f, "Incomplete opaqueData slice header (incomplete length): {}", len)
+            }
+            DepositError::UnalignedData(data) => {
+                write!(f, "Unaligned log data, expected multiple of 32 bytes, got: {}", data)
+            }
+            DepositError::FromDecode(topic) => {
+                write!(f, "Failed to decode the `from` address of the deposit log topic: {}", topic)
+            }
+            DepositError::ToDecode(topic) => {
+                write!(f, "Failed to decode the `to` address of the deposit log topic: {}", topic)
+            }
+            DepositError::InvalidOpaqueDataOffset(offset) => {
+                write!(f, "Invalid u64 opaque data content offset: {:?}", offset)
+            }
+            DepositError::InvalidOpaqueDataLength(length) => {
+                write!(f, "Invalid u64 opaque data content length: {:?}", length)
+            }
+            DepositError::OpaqueDataOverflow(data_len, opaque_len) => {
+                write!(
+                    f,
+                    "Specified opaque data length {} exceeds the deposit log event data length {}",
+                    opaque_len, data_len
+                )
+            }
+            DepositError::PaddedOpaqueDataOverflow(data_len, opaque_len) => {
+                write!(
+                    f,
+                    "Opaque data with padding exceeds the specified data length: {} > {}",
+                    opaque_len, data_len
+                )
+            }
+            DepositError::InvalidVersion(version) => {
+                write!(f, "Invalid deposit version: {}", version)
+            }
+            DepositError::UnexpectedOpaqueDataLen(len) => {
+                write!(f, "Unexpected opaque data length: {}", len)
+            }
+            DepositError::MintDecode(data) => {
+                write!(f, "Failed to decode the u128 deposit mint value: {:?}", data)
+            }
+            DepositError::GasDecode(data) => {
+                write!(f, "Failed to decode the u64 deposit gas value: {:?}", data)
+            }
+            DepositError::Custom(e) => write!(f, "Custom error: {}", e),
+        }
+    }
+}
 
 /// Source domain identifiers for deposit transactions.
 #[derive(Debug, Clone, PartialEq, Eq, Hash)]
