+++ crates/providers/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
-name = "kona-primitives"
-description = "Primitive types for kona crates"
+name = "kona-providers"
+description = "Provider traits and implementations for kona crates"
 version = "0.0.1"
 edition.workspace = true
 authors.workspace = true
@@ -14,13 +14,13 @@ tracing.workspace = true
 alloy-rlp = { workspace = true, features = ["derive"] }
 alloy-primitives = { workspace = true, features = ["rlp"] }
 
+# Local
+kona-primitives = { path = "../primitives", version = "0.0.1" }
+
 # External
-hashbrown = "0.14.3"
-unsigned-varint = "0.8.0"
-spin = { version = "0.9.8", features = ["mutex"] }
+async-trait = "0.1.77"
 
 # Alloy Types
-alloy-eips = { git = "https://github.com/alloy-rs/alloy", rev = "e3f2f07", default-features = false }
 alloy-sol-types = { version = "0.7.1", default-features = false }
 alloy-consensus = { git = "https://github.com/alloy-rs/alloy", rev = "e3f2f07", default-features = false }
 op-alloy-consensus = { git = "https://github.com/clabby/op-alloy", branch = "refcell/consensus-port", default-features = false }
@@ -30,7 +30,6 @@ serde = { version = "1.0.197", default-features = false, features = ["derive"], 
 
 [dev-dependencies]
 tokio = { version = "1.37", features = ["full"] }
-proptest = "1.4.0"
 tracing-subscriber = "0.3.18"
 serde_json = { version = "1.0.68", default-features = false }
